+++ src/anvil/__init__.py
@@ -1,6 +1,6 @@
-from .packages import  __version__, os, sys
+from .packages import __version__, os, sys
 
 if sys.argv[0].endswith('.py'):
-    from .core import *
+    from .api.molang import Math, Query, Variable
     from .api.tools import *
-    from .api.molang import Query, Variable, Math
+    from .core import *